+++ src/main/java/com/banking/backendApplication/repository/CardsRepository.java
@@ -8,6 +8,6 @@ import java.util.List;
 @Repository
 public interface CardsRepository extends CrudRepository<Cards, Long> {
 
-    List<Cards> findByCustomerId(int customerId);
+    List<Cards> findByCustomerId(Long customerId);
 
 }
