+++ src/main/java/com/banking/backendApplication/repository/LoanRepository.java
@@ -8,6 +8,6 @@ import java.util.List;
 @Repository
 public interface LoanRepository extends CrudRepository<Loans, Long> {
 
-    List<Loans> findByCustomerIdOrderByStartDtDesc(int customerId);
+    List<Loans> findByCustomerIdOrderByStartDtDesc(Long customerId);
 
 }
